Set month and day value at the same time (fixes #21)
Previously, a new Date object was created by calling the constructor without
arguments. This Date object will represent the current date and time. Then,
the code would proceed to set year, month and day of the month, in that order.

However, when accessing the app at the end of a month with more days than the
target month, the day would be out of range for the target month, leading the
month to be "corrected". Therefore, looking up a meeting in June on the last
day of May showed a date in July, instead.

Fixed by making use of the optional second argument to Date.setUTCMonth(),
which allows us to set the day value at the same time as the month value.